requirements.txt
# Core AI and Computer Vision
ultralytics>=8.3.0          # YOLO models (YOLO11n - latest version)
opencv-python>=4.8.0        # Computer vision and camera handling
face-recognition>=1.3.0     # Face recognition library (also for pet identification)
deepface>=0.0.79            # Advanced face recognition (optional, supports pet recognition)
numpy>=1.21.0               # Numerical computing
Pillow>=9.0.0               # Image processing

# Individual Pet Identification Libraries (Optional - for advanced pet recognition)
# tensorflow>=2.10.0        # For custom CNN pet identification models (heavy - 800MB+)
torch>=1.12.0               # Required by ultralytics for YOLO
torchvision>=0.13.0         # Required by ultralytics
scikit-learn>=1.1.0         # Machine learning metrics and similarity calculations

# Database
# SQLite (built-in with Python) - No additional dependencies required

# GUI Framework
# Note: Migrating from tkinter to PyQt6
# PyQt6>=6.5.0              # Modern GUI framework (target)
# tkinter is built-in        # Current GUI framework

# Telegram Bot
python-telegram-bot>=20.0  # Telegram bot integration
requests>=2.28.0           # HTTP requests for Telegram API
# asyncio                  # Async support for Telegram (built-in Python module)

# Performance Monitoring
psutil>=5.9.0              # System resource monitoring
GPUtil>=1.4.0              # GPU utilization tracking
matplotlib>=3.5.0          # Performance graphs and visualization

# Utilities
PyYAML>=6.0                # Configuration file parsing
python-dotenv>=0.19.0      # Environment variable management
colorama>=0.4.4            # Colored terminal output
tqdm>=4.64.0               # Progress bars

# Development and Testing
pytest>=7.0.0              # Testing framework
pytest-cov>=4.0.0          # Coverage testing
black>=22.0.0              # Code formatting
flake8>=5.0.0              # Code linting

# Optional GPU Acceleration
# onnx>=1.12.0              # Model optimization
# onnxruntime-gpu>=1.12.0   # GPU-accelerated inference
# Note: torch and torchvision moved to core requirements for pet identification

# System Requirements Notes:
# - Python 3.9+ (3.11 recommended)
# - Windows 11 (64-bit)
# - NVIDIA GPU with CUDA support (optional but recommended)
# - SQLite 3.x (built-in with Python)
# - 8GB+ RAM (16GB recommended)
# - 10GB+ free disk space
